INT-211: Committing code state used for the given demo
