+++ orchestrator/careplanservice/handle_taskfiller.go
@@ -90,7 +90,7 @@ func (s *Service) handleTaskFillerUpdate(task map[string]interface{}) error {
 
 	if questionnaire["url"] == "http://decor.nictiz.nl/fhir/Questionnaire/2.16.840.1.113883.2.4.3.11.60.909.26.34-2--20240902134017" {
 
-		log.Info().Msg("SubTask with PII is completed - marking primary Task (partOf value) to completed")
+		log.Info().Msg("SubTask with PII is completed - marking primary Task (partOf value) as accepted")
 
 		// Fetch task from Task.partOf
 		partOfRefs, ok := task["partOf"].([]interface{})
@@ -115,7 +115,7 @@ func (s *Service) handleTaskFillerUpdate(task map[string]interface{}) error {
 		}
 
 		// Change status to completed
-		partOfTask["status"] = "completed"
+		partOfTask["status"] = "accepted"
 
 		// Update the task in the FHIR server
 		err = s.fhirClient.Update(partOfTaskRef, partOfTask, partOfTask)
@@ -366,556 +366,4 @@ func (s *Service) partOf(task map[string]interface{}) (*string, error) {
 	}
 
 	return taskPartOf[0].Reference, nil
-}
-
-func (s *Service) getHardCodedHomeMonitoringPIIQuestionnaire() map[string]interface{} {
-	return map[string]interface{}{
-		"id":           "cps-questionnaire-patient-details",
-		"resourceType": "Questionnaire",
-		"meta": map[string]interface{}{
-			"source": "http://decor.nictiz.nl/fhir/4.0/sansa-",
-			"profile": []string{
-				SCP_TASK_PROFILE,
-				"http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-pop-exp",
-				"http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-render",
-			},
-			"tag": []map[string]interface{}{
-				{
-					"system": "http://hl7.org/fhir/FHIR-version",
-					"code":   "4.0.1",
-				},
-			},
-		},
-		"language": "nl-NL",
-		"url":      "http://decor.nictiz.nl/fhir/Questionnaire/2.16.840.1.113883.2.4.3.11.60.909.26.34-2--20240902134017",
-		"identifier": []map[string]interface{}{
-			{
-				"system": "urn:oid:2.16.840.1.113883.2.4.3.11.60.909.26.34",
-				"value":  "2",
-			},
-		},
-		"name":         "patient contactdetails",
-		"title":        "patient contactdetails",
-		"status":       "draft",
-		"experimental": false,
-		"date":         "2024-09-02T13:40:17Z",
-		"publisher":    "Medical Service Centre",
-		"effectivePeriod": map[string]interface{}{
-			"start": "2024-09-02T13:40:17Z",
-		},
-		"item": []map[string]interface{}{
-			{
-				"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2233",
-				"text":   "Naamgegevens",
-				"_text": map[string]interface{}{
-					"extension": []map[string]interface{}{
-						{
-							"extension": []map[string]interface{}{
-								{
-									"url":       "lang",
-									"valueCode": "en-US",
-								},
-								{
-									"url":         "content",
-									"valueString": "NameInformation",
-								},
-							},
-							"url": "http://hl7.org/fhir/StructureDefinition/translation",
-						},
-					},
-				},
-				"type":     "group",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-				"item": []map[string]interface{}{
-					{
-						"extension": []map[string]interface{}{
-							{
-								"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-								"valueExpression": map[string]interface{}{
-									"language":   "text/fhirpath",
-									"expression": "%patient.name.given.first()",
-								},
-							},
-						},
-						"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2234",
-						"text":   "Voornamen",
-						"_text": map[string]interface{}{
-							"extension": []map[string]interface{}{
-								{
-									"extension": []map[string]interface{}{
-										{
-											"url":       "lang",
-											"valueCode": "en-US",
-										},
-										{
-											"url":         "content",
-											"valueString": "FirstNames",
-										},
-									},
-									"url": "http://hl7.org/fhir/StructureDefinition/translation",
-								},
-							},
-						},
-						"type":     "string",
-						"required": true,
-						"repeats":  true,
-						"readOnly": false,
-					},
-					{
-						"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2238",
-						"text":   "Geslachtsnaam",
-						"_text": map[string]interface{}{
-							"extension": []map[string]interface{}{
-								{
-									"extension": []map[string]interface{}{
-										{
-											"url":       "lang",
-											"valueCode": "en-US",
-										},
-										{
-											"url":         "content",
-											"valueString": "LastName",
-										},
-									},
-									"url": "http://hl7.org/fhir/StructureDefinition/translation",
-								},
-							},
-						},
-						"type":     "group",
-						"required": true,
-						"repeats":  true,
-						"readOnly": false,
-						"item": []map[string]interface{}{
-							{
-								"extension": []map[string]interface{}{
-									{
-										"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-										"valueExpression": map[string]interface{}{
-											"language":   "text/fhirpath",
-											"expression": "%patient.name.given.last()",
-										},
-									},
-								},
-								"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2239",
-								"text":   "Voorvoegsels",
-								"_text": map[string]interface{}{
-									"extension": []map[string]interface{}{
-										{
-											"extension": []map[string]interface{}{
-												{
-													"url":       "lang",
-													"valueCode": "en-US",
-												},
-												{
-													"url":         "content",
-													"valueString": "Prefix",
-												},
-											},
-											"url": "http://hl7.org/fhir/StructureDefinition/translation",
-										},
-									},
-								},
-								"type":     "string",
-								"required": false,
-								"repeats":  false,
-								"readOnly": false,
-							},
-							{
-								"extension": []map[string]interface{}{
-									{
-										"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-										"valueExpression": map[string]interface{}{
-											"language":   "text/fhirpath",
-											"expression": "%patient.name.family",
-										},
-									},
-								},
-								"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2240",
-								"text":   "Achternaam",
-								"_text": map[string]interface{}{
-									"extension": []map[string]interface{}{
-										{
-											"extension": []map[string]interface{}{
-												{
-													"url":       "lang",
-													"valueCode": "en-US",
-												},
-												{
-													"url":         "content",
-													"valueString": "LastName",
-												},
-											},
-											"url": "http://hl7.org/fhir/StructureDefinition/translation",
-										},
-									},
-								},
-								"type":     "string",
-								"required": true,
-								"repeats":  false,
-								"readOnly": false,
-							},
-						},
-					},
-				},
-			},
-			{
-				"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2257",
-				"text":   "Contactgegevens",
-				"_text": map[string]interface{}{
-					"extension": []map[string]interface{}{
-						{
-							"extension": []map[string]interface{}{
-								{
-									"url":       "lang",
-									"valueCode": "en-US",
-								},
-								{
-									"url":         "content",
-									"valueString": "ContactInformation",
-								},
-							},
-							"url": "http://hl7.org/fhir/StructureDefinition/translation",
-						},
-					},
-				},
-				"type":     "group",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-				"item": []map[string]interface{}{
-					{
-						"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2258",
-						"text":   "Telefoonnummers",
-						"_text": map[string]interface{}{
-							"extension": []map[string]interface{}{
-								{
-									"extension": []map[string]interface{}{
-										{
-											"url":       "lang",
-											"valueCode": "en-US",
-										},
-										{
-											"url":         "content",
-											"valueString": "TelephoneNumbers",
-										},
-									},
-									"url": "http://hl7.org/fhir/StructureDefinition/translation",
-								},
-							},
-						},
-						"type":     "group",
-						"required": true,
-						"repeats":  false,
-						"readOnly": false,
-						"item": []map[string]interface{}{
-							{
-								"extension": []map[string]interface{}{
-									{
-										"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-										"valueExpression": map[string]interface{}{
-											"language":   "text/fhirpath",
-											"expression": "%patient.telecom.where(system='phone').value",
-										},
-									},
-								},
-								"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2259",
-								"text":   "Telefoonnummer",
-								"_text": map[string]interface{}{
-									"extension": []map[string]interface{}{
-										{
-											"extension": []map[string]interface{}{
-												{
-													"url":       "lang",
-													"valueCode": "en-US",
-												},
-												{
-													"url":         "content",
-													"valueString": "TelephoneNumber",
-												},
-											},
-											"url": "http://hl7.org/fhir/StructureDefinition/translation",
-										},
-									},
-								},
-								"type":     "string",
-								"required": true,
-								"repeats":  false,
-								"readOnly": false,
-							},
-						},
-					},
-					{
-						"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2263",
-						"text":   "EmailAdressen",
-						"_text": map[string]interface{}{
-							"extension": []map[string]interface{}{
-								{
-									"extension": []map[string]interface{}{
-										{
-											"url":       "lang",
-											"valueCode": "en-US",
-										},
-										{
-											"url":         "content",
-											"valueString": "EmailAddresses",
-										},
-									},
-									"url": "http://hl7.org/fhir/StructureDefinition/translation",
-								},
-							},
-						},
-						"type":     "group",
-						"required": true,
-						"repeats":  false,
-						"readOnly": false,
-						"item": []map[string]interface{}{
-							{
-								"extension": []map[string]interface{}{
-									{
-										"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-										"valueExpression": map[string]interface{}{
-											"language":   "text/fhirpath",
-											"expression": "%patient.telecom.where(system='email').value",
-										},
-									},
-								},
-								"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2264",
-								"text":   "EmailAdres",
-								"_text": map[string]interface{}{
-									"extension": []map[string]interface{}{
-										{
-											"extension": []map[string]interface{}{
-												{
-													"url":       "lang",
-													"valueCode": "en-US",
-												},
-												{
-													"url":         "content",
-													"valueString": "EmailAddress",
-												},
-											},
-											"url": "http://hl7.org/fhir/StructureDefinition/translation",
-										},
-									},
-								},
-								"type":     "string",
-								"required": true,
-								"repeats":  false,
-								"readOnly": false,
-							},
-						},
-					},
-				},
-			},
-			{
-				"extension": []map[string]interface{}{
-					{
-						"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-						"valueExpression": map[string]interface{}{
-							"language":   "text/fhirpath",
-							"expression": "%patient.identifier.where(system='http://fhir.nl/fhir/NamingSystem/bsn').value",
-						},
-					},
-				},
-				"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2266",
-				"text":   "Burgerservicenummer (OID: 2.16.840.1.113883.2.4.6.3)",
-				"_text": map[string]interface{}{
-					"extension": []map[string]interface{}{
-						{
-							"extension": []map[string]interface{}{
-								{
-									"url":       "lang",
-									"valueCode": "en-US",
-								},
-								{
-									"url":         "content",
-									"valueString": "Burgerservicenummer (OID: 2.16.840.1.113883.2.4.6.3)",
-								},
-							},
-							"url": "http://hl7.org/fhir/StructureDefinition/translation",
-						},
-					},
-				},
-				"type":     "string",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-			{
-				"extension": []map[string]interface{}{
-					{
-						"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-						"valueExpression": map[string]interface{}{
-							"language":   "text/fhirpath",
-							"expression": "%patient.birthDate",
-						},
-					},
-				},
-				"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2267",
-				"text":   "Geboortedatum",
-				"_text": map[string]interface{}{
-					"extension": []map[string]interface{}{
-						{
-							"extension": []map[string]interface{}{
-								{
-									"url":       "lang",
-									"valueCode": "en-US",
-								},
-								{
-									"url":         "content",
-									"valueString": "DateOfBirth",
-								},
-							},
-							"url": "http://hl7.org/fhir/StructureDefinition/translation",
-						},
-					},
-				},
-				"type":     "dateTime",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-			{
-				"extension": []map[string]interface{}{
-					{
-						"url": "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression",
-						"valueExpression": map[string]interface{}{
-							"language":   "text/fhirpath",
-							"expression": "%patient.gender",
-						},
-					},
-				},
-				"linkId": "2.16.840.1.113883.2.4.3.11.60.909.2.2.2268",
-				"text":   "Geslacht",
-				"_text": map[string]interface{}{
-					"extension": []map[string]interface{}{
-						{
-							"extension": []map[string]interface{}{
-								{
-									"url":       "lang",
-									"valueCode": "en-US",
-								},
-								{
-									"url":         "content",
-									"valueString": "Gender",
-								},
-							},
-							"url": "http://hl7.org/fhir/StructureDefinition/translation",
-						},
-					},
-				},
-				"type":     "choice",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-				"answerOption": []map[string]interface{}{
-					{
-						"valueCoding": map[string]interface{}{
-							"system":  "http://hl7.org/fhir/administrative-gender",
-							"code":    "other",
-							"display": "Other",
-						},
-					},
-					{
-						"valueCoding": map[string]interface{}{
-							"system":  "http://hl7.org/fhir/administrative-gender",
-							"code":    "male",
-							"display": "Male",
-						},
-					},
-					{
-						"valueCoding": map[string]interface{}{
-							"system":  "http://hl7.org/fhir/administrative-gender",
-							"code":    "female",
-							"display": "Female",
-						},
-					},
-					{
-						"valueCoding": map[string]interface{}{
-							"system":  "http://hl7.org/fhir/administrative-gender",
-							"code":    "unknown",
-							"display": "Unknown",
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func (s *Service) getHardCodedHomeMonitoringQuestionnaire() map[string]interface{} {
-	return map[string]interface{}{
-		"id":           "cps-questionnaire-telemonitoring-enrollment-criteria",
-		"resourceType": "Questionnaire",
-		"meta": map[string]interface{}{
-			"source": "http://decor.nictiz.nl/fhir/4.0/sansa-",
-			"tag": []map[string]interface{}{
-				{
-					"system": "http://hl7.org/fhir/FHIR-version",
-					"code":   "4.0.1",
-				},
-			},
-		},
-		"language": "nl-NL",
-		"url":      "http://decor.nictiz.nl/fhir/Questionnaire/2.16.840.1.113883.2.4.3.11.60.909.26.34-1--20240902134017",
-		"identifier": []map[string]interface{}{
-			{
-				"system": "urn:ietf:rfc:3986",
-				"value":  "urn:oid:2.16.840.1.113883.2.4.3.11.60.909.26.34-1",
-			},
-		},
-		"name":         "Telemonitoring - enrollment criteria",
-		"title":        "Telemonitoring - enrollment criteria",
-		"status":       "active",
-		"experimental": false,
-		"date":         "2024-09-02T13:40:17Z",
-		"publisher":    "Medical Service Centre",
-		"effectivePeriod": map[string]interface{}{
-			"start": "2024-09-02T13:40:17Z",
-		},
-		"item": []map[string]interface{}{
-			{
-				"linkId":   "2.16.840.1.113883.2.4.3.11.60.909.2.2.2208",
-				"text":     "Patient heeft smartphone",
-				"type":     "boolean",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-			{
-				"linkId":   "2.16.840.1.113883.2.4.3.11.60.909.2.2.2209",
-				"text":     "Patient of mantelzorger leest e-mail op smartphone",
-				"type":     "boolean",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-			{
-				"linkId":   "2.16.840.1.113883.2.4.3.11.60.909.2.2.2210",
-				"text":     "Patient of mantelzorger kan apps installeren op smartphone",
-				"type":     "boolean",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-			{
-				"linkId":   "2.16.840.1.113883.2.4.3.11.60.909.2.2.2211",
-				"text":     "Patient of mantelzorger is Nederlandse taal machtig",
-				"type":     "boolean",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-			{
-				"linkId":   "2.16.840.1.113883.2.4.3.11.60.909.2.2.2212",
-				"text":     "Patient beschikt over een weegschaal of bloeddrukmeter (of gaat deze aanschaffen)",
-				"type":     "boolean",
-				"required": true,
-				"repeats":  false,
-				"readOnly": false,
-			},
-		},
-	}
 }
